requirements.txt
# Core dependencies
numpy==2.4.2
pandas==3.0.0
requests==2.32.5
pyyaml==6.0.3

# Machine learning
tensorflow>=2.15.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0

# API & Automation (optional - for n8n integration)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
